dd: get fd on newly created element
Define a function that enables libreport users to create a new element
without the need to pass its contents. Sometimes we need to create
element as a privileged user but write data to it as an unprivileged
user.

TODO: check for locked dump dir
TODO: add tests
TODO: better documentation
TODO: shall we track these fds? should we close them?

Signed-off-by: Jakub Filak <jfilak@redhat.com>