dd: detect chroot environment from saved data
If a problem comes from a chroot environment, libreport gathers OS
details from the chrooted root directory. However, this step is
performed only if we call dd_create_basic_files with the chroot
argument. This makes it hard to use from high level ABRT APIs.

This patch adds the possibility to pass the chroot path via dump dir
elements. So the feature will be easily available to all kind of APIs.
The only requirement will be to set FILENAME_ROOTDIR before calling the
function.

Signed-off-by: Jakub Filak <jfilak@redhat.com>